scripts/github-backup/requirements.txt
GitPython==3.1.31
PyGithub==1.58.2
python-dotenv==1.0.0
